+++ packages/categories/server/places.js
@@ -1,4 +1,28 @@
 
+Places.after.update(function(userId, doc, fieldNames, modifier, options) {
+
+	if(_.contains(fieldNames,'cats') && modifier['$addToSet']) {
+
+		var catData = {},
+			cats = modifier.$addToSet.cats.$each;
+
+		_.each(cats, function(cat) {
+
+			catData = _.extend({}, K.schemas.cat, {
+				name: cat,
+				type: 'place'//user, place, all
+			});
+
+			Categories.upsert({name: catData.name}, {
+				$set: catData
+			});
+		});
+
+		console.log('Cats: after update', cats);
+	}
+
+});
+
 Meteor.publish('placesByCategory', function(cat) {
 	if(this.userId) {
 		var cur = K.findPlacesByCategory(cat);